+++ app/Http/Controllers/OfficeController.php
@@ -5,7 +5,10 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Office;
 use App\Models\TotalDemandPrint;
+use Illuminate\Support\Facades\Redirect;
 use PDF;
+use Illuminate\Support\Facades\Auth;
+
 
 class OfficeController extends Controller
 {
@@ -14,9 +17,10 @@ class OfficeController extends Controller
     | AUTHENTICATION
     |-----------------------------------------
     */
-    public function __construct(){
-    	// body
-    	$this->middleware('auth');
+    public function __construct()
+    {
+        // body
+        $this->middleware('auth');
     }
 
     /*
@@ -24,21 +28,21 @@ class OfficeController extends Controller
     | SHOW VIEW INDEX
     |-----------------------------------------
     */
-    public function index(Request $request){
+    public function index(Request $request)
+    {
         // body
         $grandTotal = 1000;
-        if($request->has('search_keywords')){
+        if ($request->has('search_keywords')) {
 
             $search_keywords = $request->search_keywords;
             $offices = Office::where('cadastral_zone', 'LIKE', "%$search_keywords%")
-            ->orWhere('asset_no', 'LIKE', "%$search_keywords%")
-            ->orWhere('prop_addr', 'LIKE', "%$search_keywords%")
-            ->orWhere('pid', 'LIKE', "%$search_keywords%")
-            ->orderBy('pid', 'DESC')
-            ->paginate(20);
-
-        }else{
-            $offices = Office::where('grand_total', '<',$grandTotal)->sortable('pid', 'DESC')->paginate(20);
+                ->orWhere('asset_no', 'LIKE', "%$search_keywords%")
+                ->orWhere('prop_addr', 'LIKE', "%$search_keywords%")
+                ->orWhere('pid', 'LIKE', "%$search_keywords%")
+                ->orderBy('pid', 'DESC')
+                ->paginate(20);
+        } else {
+            $offices = Office::where('grand_total', '<', $grandTotal)->sortable('pid', 'DESC')->paginate(20);
         }
 
         return view('office.index', compact('offices'));
@@ -49,13 +53,20 @@ class OfficeController extends Controller
     | SHOW VIEW INDEX
     |-----------------------------------------
     */
-    public function view(Request $request){
+    public function view(Request $request)
+    {
         // body
-        $office = Office::where('pid', $request->pid)->first();
-        return view('office.show', compact('office'));
+        if (Auth::user()->user_type == 'super') {
+            $office = Office::where('pid', $request->pid)->first();
+            return view('office.show', compact('office'));
+        } else {
+            $msg= 'you are not allow to view this page';
+            return Redirect::back()->with($msg);
+        }
     }
 
-    public function search(Request $request){
+    public function search(Request $request)
+    {
 
         $powerMeter = Office::query();
 
@@ -64,30 +75,30 @@ class OfficeController extends Controller
         $asset_no                 = $request->asset_no;
         $prop_addr                = $request->prop_addr;
 
-        if($request->has('keywords')){
-            $powerMeter->where(function($query) use ($request) {
-            $query->where('pid', 'LIKE', "%{$pid}%")
-            ->orWhere('cadastral_zone', 'LIKE', "%{$cadastral_zone}%")
-            ->orWhere('asset_no', 'LIKE', "%{$asset_no}%")
-            ->orWhere('prop_addr', 'LIKE', "%{$prop_addr}%");
+        if ($request->has('keywords')) {
+            $powerMeter->where(function ($query) use ($request) {
+                $query->where('pid', 'LIKE', "%{$pid}%")
+                    ->orWhere('cadastral_zone', 'LIKE', "%{$cadastral_zone}%")
+                    ->orWhere('asset_no', 'LIKE', "%{$asset_no}%")
+                    ->orWhere('prop_addr', 'LIKE', "%{$prop_addr}%");
             });
         }
 
         return view('office.index', compact('office'));
     }
 
-     // Generate PDF
+    // Generate PDF
     public function createPDF()
     {
-      // retreive all records from db
-      $office = Office::all();
+        // retreive all records from db
+        $office = Office::all();
 
-      // share data to view
-      view()->share('office',$office);
-      $pdf = PDF::loadView('pdf_view', $data);
+        // share data to view
+        view()->share('office', $office);
+        $pdf = PDF::loadView('pdf_view', $data);
 
-      // download PDF file with download method
-      return $pdf->download('pdf_file.pdf');
+        // download PDF file with download method
+        return $pdf->download('pdf_file.pdf');
     }
 
     public function create()
@@ -95,39 +106,46 @@ class OfficeController extends Controller
         return view('office.create');
     }
 
-    public function previewAll(Request $request){
+    public function previewAll(Request $request)
+    {
         // body
-        $office_ids = json_decode($request->office_ids);
+        if (Auth::user()->user_type == 'super') {
+            $office_ids = json_decode($request->office_ids);
 
-        $offices = Office::whereIn('id', $office_ids)->orderBy('pid', 'DESC')->get();
-        return view('office.preview', compact('offices'));
+            $offices = Office::whereIn('id', $office_ids)->orderBy('pid', 'DESC')->get();
+            return view('office.preview', compact('offices'));
+        } else {
+            // return 'you are not allow to view this page';
+            return redirect()->back(); 
+        }
     }
 
 
-    public function store(Request $request){
+    public function store(Request $request)
+    {
 
         $office = new Office();
-        $office->pid                =$request->input('pid');
-        $office->occupant           =$request->input('occupant');
-        $office->prop_addr          =$request->input('prop_addr');
-        $office->street_name        =$request->input('street_name');
-        $office->asset_no           =$request->input('asset_no');
-        $office->cadastral_zone     =$request->input('cadastral_zone');
-        $office->prop_type          =$request->input('prop_type');
-        $office->prop_use           =$request->input('prop_use');
-        $office->rating_dist        =$request->input('rating_dist');
-        $office->annual_value       =$request->input('annual_value');
-        $office->rate_payable       =0.04 * $office->annual_value;
-        $office->arrears            =$request->input('arrears');
-        $office->penalty            =$request->input('penalty');
-        $office->paid_amount        =$request->input('paid_amount');
+        $office->pid                = $request->input('pid');
+        $office->occupant           = $request->input('occupant');
+        $office->prop_addr          = $request->input('prop_addr');
+        $office->street_name        = $request->input('street_name');
+        $office->asset_no           = $request->input('asset_no');
+        $office->cadastral_zone     = $request->input('cadastral_zone');
+        $office->prop_type          = $request->input('prop_type');
+        $office->prop_use           = $request->input('prop_use');
+        $office->rating_dist        = $request->input('rating_dist');
+        $office->annual_value       = $request->input('annual_value');
+        $office->rate_payable       = 0.04 * $office->annual_value;
+        $office->arrears            = $request->input('arrears');
+        $office->penalty            = $request->input('penalty');
+        $office->paid_amount        = $request->input('paid_amount');
         $office->grand_total        = $office->rate_payable + $office->arrears + $office->penalty;
-        $office->category           =$request->input('category');
-        $office->group              =$request->input('group');
-        $office->active             =$request->input('active');
+        $office->category           = $request->input('category');
+        $office->group              = $request->input('group');
+        $office->active             = $request->input('active');
 
         $office->save();
-        return redirect()->route('offices')->with('success','office information has been created Successfully');
+        return redirect()->route('offices')->with('success', 'office information has been created Successfully');
     }
 
 
@@ -137,46 +155,51 @@ class OfficeController extends Controller
 
     public function edit($id)
     {
-
-        $office = Office::findOrFail($id);
-        return view('office.edit',compact('office'));
+        if (Auth::user()->user_type == 'super') {
+            $office = Office::findOrFail($id);
+            return view('office.edit', compact('office'));
+        } else {
+            // return 'you are not allow to view this page';
+            return redirect()->back(); 
+        }
     }
 
     // |-----------------------------------------
     // | MODIFY or UPDATE DATA
     // |-----------------------------------------
 
-    public function update($id, Request $request){
+    public function update($id, Request $request)
+    {
 
 
+        if (Auth::user()->user_type == 'super') {
+            $office = Office::find($id);
+            $office->pid                = $request->pid;
+            $office->occupant           = $request->occupant;
+            $office->prop_addr          = $request->prop_addr;
+            $office->street_name        = $request->street_name;
+            $office->asset_no           = $request->asset_no;
+            $office->cadastral_zone     = $request->cadastral_zone;
+            $office->prop_type          = $request->prop_type;
+            $office->prop_use           = $request->prop_use;
+            $office->rating_dist        = $request->rating_dist;
+            $office->annual_value       = $request->annual_value;
+            $office->rate_payable       = 0.04 * $office->annual_value;
+            $office->arrears            = $request->arrears;
+            $office->penalty            = $request->penalty;
+            $office->paid_amount        = $request->paid_amount;
+            $office->grand_total        = $office->rate_payable + $office->arrears + $office->penalty - $office->paid_amount;
+            $office->category           = $request->category;
+            $office->group              = $request->group;
+            $office->active             = $request->active;
 
-        $office = Office::find($id);
+            $office->update();
 
-
-
-
-        $office->pid                =$request->pid;
-        $office->occupant           =$request->occupant;
-        $office->prop_addr          =$request->prop_addr;
-        $office->street_name        =$request->street_name;
-        $office->asset_no           =$request->asset_no;
-        $office->cadastral_zone     =$request->cadastral_zone;
-        $office->prop_type          =$request->prop_type;
-        $office->prop_use           =$request->prop_use;
-        $office->rating_dist        =$request->rating_dist;
-        $office->annual_value       =$request->annual_value;
-        $office->rate_payable       = 0.04 * $office->annual_value;
-        $office->arrears            =$request->arrears;
-        $office->penalty            =$request->penalty;
-        $office->paid_amount        =$request->paid_amount;
-        $office->grand_total        =$office->rate_payable + $office->arrears + $office->penalty - $office->paid_amount ;
-        $office->category           =$request->category;
-        $office->group              =$request->group;
-        $office->active             =$request->active;
-
-        $office->update();
-
-        return redirect()->route('offices')->with('success','office information has been created Successfully');
+            return redirect()->route('offices')->with('success', 'office information has been created Successfully');
+        } else {
+            // return 'you are not allow to view this page';
+            return redirect()->back(); 
+        }
     }
 
     // /*
@@ -186,17 +209,26 @@ class OfficeController extends Controller
 
     public function destroy($id)
     {
-
-        $office = Office::find($id);
-        $office->delete();
-        return view('office.index',compact('office'));
+        if (Auth::user()->user_type == 'super') {
+            $office = Office::find($id);
+            $office->delete();
+            return view('office.index', compact('office'));
+        } else {
+            // return 'you are not allow to perform this action';
+            return redirect()->back(); 
+        }
     }
 
-    public function saveTotalPrint(Request $request )
+    public function saveTotalPrint(Request $request)
     {
+        if (Auth::user()->user_type == 'super') {
 
-        $total_demand_print = new TotalDemandPrint();
-        $total_demand_print->addOne($request);
+            $total_demand_print = new TotalDemandPrint();
+            $total_demand_print->addOne($request);
+        } else {
+            // return 'you are not allow to perform this action';
+            return redirect()->back(); 
+        }
     }
 
     public function sortSearch(Request $request)
@@ -204,26 +236,23 @@ class OfficeController extends Controller
 
 
         $sortSearch =  Office::where('pid', 'like', '%' . $request->pid . '%')
-        ->where('asset_no', 'like', '%' . $request->asset_no . '%')
-        ->where('street_name', 'like', '%' . $request->street_name . '%')
-        ->get();
+            ->where('asset_no', 'like', '%' . $request->asset_no . '%')
+            ->where('street_name', 'like', '%' . $request->street_name . '%')
+            ->get();
 
 
-    //     if($request->has('sortSearch')){
+        //     if($request->has('sortSearch')){
 
-    //         $sortSearch = $request->sortSearch;
-    //         $offices = Office::where('pid', 'like', '%' . $request->pid . '%')
-    //                ->where('asset_no', 'like', '%' . $request->asset_no . '%')
-    //                ->where('street_name', 'like', '%' . $request->street_name . '%')
-    //                ->orderBy('pid', 'DESC')
-    //                ->paginate(10);
+        //         $sortSearch = $request->sortSearch;
+        //         $offices = Office::where('pid', 'like', '%' . $request->pid . '%')
+        //                ->where('asset_no', 'like', '%' . $request->asset_no . '%')
+        //                ->where('street_name', 'like', '%' . $request->street_name . '%')
+        //                ->orderBy('pid', 'DESC')
+        //                ->paginate(10);
 
-    // }else{
-    //     $offices = Office::orderBy('pid', 'DESC')->paginate(20);
+        // }else{
+        //     $offices = Office::orderBy('pid', 'DESC')->paginate(20);
 
-    //     }return view('office.index', compact('offices'));
+        //     }return view('office.index', compact('offices'));
     }
-
-
-
 }
